Adding Unequal Variance to Signature Analysis T-Test (#186)
* make independent t-test have equal variance!

* add tissue balance and rerun separate notebook

* update top features in notebook 2

* scientific data updated the URL!

* add the actual data in case it updates again

* make sure equal variance is false

* remove hard coded best features in visualization

also calculate t-test in figures after filtering tissues

* update full separation plot figure

* rerun full signature pipeline

* make points in figure bigger b/c using test set

* rerun pipeline

* closes #187

* modify figure for rerunning analysis

* rerun pipeline

+++ 10.gene-expression-signatures/scripts/nbconverted/0.download-validation-data.py
@@ -53,7 +53,7 @@ get_ipython().system(' md5sum "download/2019-01-22-CellLineSTAR-fpkm-2pass_matri
 # In[6]:
 
 
-url = "https://www.nature.com/articles/sdata201733/tables/3"
+url = "https://www.nature.com/articles/sdata201733/tables/4"
 name = "nbl_cellline_phenotype.txt"
 path = os.path.join("download", name)
 
@@ -61,12 +61,16 @@ path = os.path.join("download", name)
 # In[7]:
 
 
-html = requests.get(url).content
+if not os.path.isfile(path):
+    html = requests.get(url).content
 
-pheno_df = pd.read_html(html)[0]
-pheno_df['Cell Line'] = pheno_df['Cell Line'].str.replace("-", "")
+    pheno_df = pd.read_html(html)[0]
+    pheno_df['Cell Line'] = pheno_df['Cell Line'].str.replace("-", "")
 
-pheno_df.to_csv(path, sep='\t', index=False)
+    pheno_df.to_csv(path, sep='\t', index=False)
+
+else:
+    pheno_df = pd.read_csv(path, sep="\t")
 
 pheno_df.head()
